+++ packages/client/src/pages/EmissionsMetricsPage/EmissionsFilterBar/Filters/CloudProviderFilter/CloudProviderFilter.tsx
@@ -4,9 +4,8 @@
 
 import React, { FunctionComponent } from 'react'
 import { CLOUD_PROVIDER_OPTIONS } from '../../utils/DropdownConstants'
-import { FilterProps } from '../../utils/Filters'
 import DropdownFilter from '../DropdownFilter'
-import { DropdownOption } from 'Types'
+import { DropdownOption, FilterProps } from 'Types'
 
 const CloudProviderFilter: FunctionComponent<FilterProps> = ({
   filters,