+++ packages/core/src/unknown/UnknownEstimator.ts
@@ -2,23 +2,20 @@
  * © 2021 Thoughtworks, Inc.
  */
 
-import {
+import { containsAny } from '@cloud-carbon-footprint/common'
+import CloudConstants, {
   CloudConstantsEmissionsFactors,
-  CloudConstants,
-  FootprintEstimate,
-  IFootprintEstimator,
+} from '../CloudConstantsTypes'
+import FootprintEstimate, {
   Co2ePerCost,
   estimateKwh,
   EstimateClassification,
-} from '../.'
-import { UnknownUsage } from '.'
+} from '../FootprintEstimate'
+import IFootprintEstimator from '../IFootprintEstimator'
+import UnknownUsage from './UnknownUsage'
 
 export default class UnknownEstimator implements IFootprintEstimator {
-  unknownUsageTypes: { [key: string]: string }
-
-  constructor(unknownUsageTypes: { [key: string]: string }) {
-    this.unknownUsageTypes = unknownUsageTypes
-  }
+  constructor(private unknownUsageTypesMapping: { [key: string]: string[] }) {}
 
   estimate(
     data: UnknownUsage[],
@@ -28,9 +25,7 @@ export default class UnknownEstimator implements IFootprintEstimator {
   ): FootprintEstimate[] {
     return data.map((data: UnknownUsage) => {
       // consider adding a console error to add unknown usageUnit to map
-      const classification = this.unknownUsageTypes[data.usageUnit]
-        ? this.unknownUsageTypes[data.usageUnit]
-        : EstimateClassification.UNKNOWN
+      const classification = this.getClassification(data)
       const usesAverageCPUConstant =
         classification === EstimateClassification.COMPUTE
       const estimatedCO2Emissions = this.estimateCo2(
@@ -51,6 +46,23 @@ export default class UnknownEstimator implements IFootprintEstimator {
       }
     })
   }
+
+  private getClassification(data: UnknownUsage) {
+    if (
+      this.unknownUsageTypesMapping[data.usageUnit]?.[1] ===
+      EstimateClassification.MEMORY
+    ) {
+      if (containsAny(['Memory'], data.usageType)) {
+        return EstimateClassification.MEMORY
+      }
+      return EstimateClassification.STORAGE
+    }
+
+    return this.unknownUsageTypesMapping[data.usageUnit]?.[0]
+      ? this.unknownUsageTypesMapping[data.usageUnit][0]
+      : EstimateClassification.UNKNOWN
+  }
+
   private estimateCo2(
     cost: number,
     co2ePerCost: Co2ePerCost,
@@ -59,7 +71,10 @@ export default class UnknownEstimator implements IFootprintEstimator {
     // This creates a coefficient based on the co2e per cost ratio of a given usage classification,
     // then multiplies the coefficient by the reclassified unknown usage cost to get estimated co2e.
     // If the new classification is still unknown, the coefficient is the average from the totals.
-    if (classification === EstimateClassification.UNKNOWN)
+    if (
+      classification === EstimateClassification.UNKNOWN ||
+      !co2ePerCost[classification].cost
+    )
       return (co2ePerCost.total.co2e / co2ePerCost.total.cost) * cost
 
     return (