+++ packages/core/src/unknown/UnknownUsage.ts
@@ -2,9 +2,10 @@
  * © 2021 Thoughtworks, Inc.
  */
 
-import { IUsageData } from '../.'
+import IUsageData from '../IUsageData'
 
 export default interface UnknownUsage extends IUsageData {
   readonly cost: number
   readonly usageUnit: string
+  readonly usageType?: string
 }