+++ packages/gcp/src/application/GCPAccount.ts
@@ -16,6 +16,7 @@ import {
   StorageEstimator,
   NetworkingEstimator,
   MemoryEstimator,
+  UnknownEstimator,
   CloudProviderAccount,
 } from '@cloud-carbon-footprint/core'
 import {
@@ -30,6 +31,7 @@ import {
   GCP_CLOUD_CONSTANTS,
   GCP_EMISSIONS_FACTORS_METRIC_TON_PER_KWH,
 } from '../domain'
+import { UNKNOWN_USAGE_TO_ASSUMED_USAGE_MAPPING } from '../lib/BillingExportTypes'
 
 export default class GCPAccount extends CloudProviderAccount {
   constructor(
@@ -76,6 +78,7 @@ export default class GCPAccount extends CloudProviderAccount {
       new StorageEstimator(GCP_CLOUD_CONSTANTS.HDDCOEFFICIENT),
       new NetworkingEstimator(GCP_CLOUD_CONSTANTS.NETWORKING_COEFFICIENT),
       new MemoryEstimator(GCP_CLOUD_CONSTANTS.MEMORY_COEFFICIENT),
+      new UnknownEstimator(UNKNOWN_USAGE_TO_ASSUMED_USAGE_MAPPING),
       new BigQuery({ projectId: this.id }),
     )
     return billingExportTableService.getEstimates(startDate, endDate)